requirements.txt
six>=1.14.0
tqdm>=4.46.0
greenlet>=0.4.15
PyJWT>=1.7.1
PyMySQL>=0.9.3
python-memcached>=1.59
tormysql>=0.4.2
tornado>=6.0.4
setuptools>=46.1.3
qrcode>=6.1